python/services/reachy-exporter/pyproject.toml
[project]
name = "reachy-exporter"
version = "0.1.0"
description = "Edge telemetry exporter for Reachy-Mini hardware"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
  "streaming-agents-core",
  "httpx>=0.27",
  "boto3>=1.35",
  "python-ulid>=3.0",
]

[project.optional-dependencies]
imu = ["reachy-mini"]

[project.scripts]
reachy-exporter = "reachy_exporter.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reachy_exporter"]

[tool.uv.sources]
streaming-agents-core = { workspace = true }
